requirements.txt
Flask==2.0.1
flask-cors==3.1.1
pytesseract==0.3.8
Pillow==8.2.0
nltk==3.6.3
sklearn==0.24.2
joblib==1.0.1
googletrans==4.0.0-rc1
